requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
discord.py==1.7.3
feedparser==6.0.8
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==5.2.0
numpy==1.21.5
pycparser==2.21
PyNaCl==1.4.0
python-dotenv==0.19.2
pytube==11.0.2
rfc3986==1.5.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
typing-extensions==4.0.1
yarl==1.7.2